04-pdf_analyzer/requirements.txt
Flask
pdfplumber
pdfkit
python-dateutil
